auth-service/requirements.txt
django==5.1.5
djangorestframework==3.15.2
PyJWT==2.10.1
